# Optimize insertion sort