@nagisa 2017-03-25
Since this is internal to libstd/core, could you check whether the inclusive range syntax makes things better as well?

i.e. `2 ... v.len()`