@stjepang 2017-03-25
@nagisa Inclusive range syntax makes performance slightly worse, actually...

With the old insertion sort and with inclusive range syntax there's a bound check at the beginning of insertion sort. This PR removes the bound check.

If you want to play with this, here's a [playpen link](https://play.rust-lang.org/?gist=e6c247daa566155657b5045ead6a0d91&version=nightly&backtrace=0).